+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## NEXT
 
-  *
+  * Add sorting comparator spaceship (&lt;=&gt;) to each of the models.
 
 ## v1.0.13 (2013-08-13)